+++ src/demo-app/features/dialog/dialog-demo.ts
@@ -14,7 +14,7 @@ export class DialogDemo {
     message: 'Do you want to <b>logout</b>?',
     okButtonText: 'Yes',
     cancelButtonText: 'No',
-
+    loading: true,
   };
   dialogFromComponentConfig: NwbDialogConfig = {
     title: 'From component',
@@ -22,8 +22,16 @@ export class DialogDemo {
     cancelButtonText: 'No',
     width: '900px',
   };
+  dialogFromComponentConfigWithSpinner: NwbDialogConfig = {
+    title: 'From component with spinner',
+    okButtonText: 'Yes',
+    cancelButtonText: 'No',
+    width: '900px',
+    loading: true
+  };
 
   fakeComponentValue = '';
+  fakeComponent2Value = '';
   numTemplateOpens = 0;
 
   @ViewChild('dialogTemplateRef') dialogTemplateRef: TemplateRef<any>;
@@ -36,6 +44,15 @@ export class DialogDemo {
 
 
     let dialog = this.nwbDialog.open(this.dialogConfig);
+
+    dialog.ready
+      .subscribe(() => {
+        console.log('openDialog is ready');
+        setTimeout(() => {
+          console.log('openDialog is loaded');
+          dialog.hideSpinner();
+        }, 3000);
+      });
 
     dialog.config.okHandler = () => {
       dialog.disableButtonsAndMakeOkButtonLoading();
@@ -54,6 +71,31 @@ export class DialogDemo {
   openDialogFromComponent() {
     const dialog = this.nwbDialog
       .openFromComponent(FakeDialogDemoComponent, this.dialogFromComponentConfig);
+
+    dialog.componentInstance.myInput.nativeElement.value = 'Random text';
+
+    dialog
+      .afterClosed()
+      .subscribe(fromOkButton => {
+        if (fromOkButton) {
+          this.fakeComponentValue = dialog.componentInstance.myInput.nativeElement.value;
+        }
+      });
+  }
+
+  openDialogFromComponentWithSpinner() {
+    const dialog = this.nwbDialog
+      .openFromComponent(FakeDialogDemoComponent, this.dialogFromComponentConfigWithSpinner);
+
+    dialog.ready
+      .subscribe(() => {
+        console.log('openDialogFromComponentWithSpinner is ready');
+        dialog.componentInstance.loading
+          .subscribe(() => {
+            console.log('FakeDialogDemoComponent is loaded');
+            dialog.hideSpinner();
+          });
+      });
 
     dialog.componentInstance.myInput.nativeElement.value = 'Random text';