+++ src/demo-app/features/spinner/spinner-demo.ts
@@ -1,0 +1,18 @@
+import {Component} from '@angular/core';
+
+@Component({
+  moduleId: module.id,
+  providers: [],
+  templateUrl: './spinner-demo.html',
+})
+export class SpinnerDemo {
+
+  customColor = '#f49a00';
+
+  constructor() {
+
+  }
+
+
+}
+