+++ src/lib/dialog/index.ts
@@ -2,9 +2,10 @@ import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {NwbDialogService} from './dialog.service';
 import {NwbDialogComponent} from './dialog.component';
+import {NwbSpinnerModule} from '../spinner/index';
 
 @NgModule({
-  imports: [CommonModule],
+  imports: [CommonModule, NwbSpinnerModule],
   providers: [NwbDialogService],
   declarations: [NwbDialogComponent],
   entryComponents: [NwbDialogComponent],