+++ src/lib/index.ts
@@ -3,3 +3,4 @@ export * from './module';
 export * from './dialog/index';
 export * from './snackbar/index';
 export * from './paginator/index';
+export * from './spinner/index';