+++ src/lib/module.ts
@@ -4,13 +4,14 @@ import {NwbDialogModule} from './dialog/index';
 import {OVERLAY_CONTAINER_PROVIDER} from './shared/overlay/overlay-container';
 import {DOM_SERVICE_PROVIDER} from './shared/dom/dom.service';
 import {NwbPaginatorModule} from './paginator/index';
+import {NwbSpinnerModule} from './spinner/index';
 
 
 @NgModule({
-  imports: [NwbSnackbarModule, NwbDialogModule, NwbPaginatorModule],
+  imports: [NwbSnackbarModule, NwbDialogModule, NwbPaginatorModule, NwbSpinnerModule],
   providers: [OVERLAY_CONTAINER_PROVIDER, DOM_SERVICE_PROVIDER],
   declarations: [],
-  exports: [NwbSnackbarModule, NwbDialogModule, NwbPaginatorModule],
+  exports: [NwbSnackbarModule, NwbDialogModule, NwbPaginatorModule, NwbSpinnerModule],
 })
 export class NwbModule {
 }